refactor: changed mentors list for sc
