+++ src/components/MentorList/index.tsx
@@ -1,29 +1,15 @@
-'use client';
-
-import { useMemo } from 'react';
-import { useRouter, useSearchParams } from 'next/navigation';
-import type { Topic } from '@/lib/types';
+import { use, useMemo } from 'react';
+import SelectTopic from './SelectTopic';
+import MentorCard from '../MentorCard';
+import { getAllMentors, getMentoringTopics } from '@/lib/api.server';
 
 interface MentorListProps {
-  topics: Topic[];
-  children: React.ReactNode;
+  speciality?: string;
 }
 
-const MentorList: React.FC<MentorListProps> = ({ topics, children }) => {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-
-  const speciality = searchParams.get('especialidad');
-
-  const queryTopic = (topic: string) => {
-    const url = new URL(window.location.href);
-
-    const params = new URLSearchParams(url.search);
-    params.set('especialidad', topic);
-    url.search = `?${params}`;
-
-    router.replace(url.toString());
-  };
+const MentorList: React.FC<MentorListProps> = ({ speciality }) => {
+  const topics = use(getMentoringTopics());
+  const mentors = use(getAllMentors());
 
   /**
    * sort modifies the original array, so make a copy to be safe
@@ -33,6 +19,22 @@ const MentorList: React.FC<MentorListProps> = ({ topics, children }) => {
     [topics],
   );
 
+  /**
+   * All active mentors should be first
+   */
+  const sortedMentors = useMemo(
+    () => [...mentors].sort((mentor) => (mentor.status === 'ACTIVE' ? -1 : 1)),
+    [mentors],
+  );
+
+  const filteredMentors = useMemo(() => {
+    if (!speciality) return sortedMentors;
+
+    return sortedMentors.filter((mentor) =>
+      mentor.topics?.some((topic) => topic._ref === speciality),
+    );
+  }, [sortedMentors, speciality]);
+
   return (
     <div>
       <div className="flex justify-between">
@@ -41,11 +43,7 @@ const MentorList: React.FC<MentorListProps> = ({ topics, children }) => {
         </h1>
       </div>
       <div className="relative inline-block w-full mb-6 md:w-1/2 lg:w-1/3">
-        <select
-          aria-label="Buscar"
-          onChange={(event) => queryTopic(event.target.value)}
-          className="block w-full px-4 py-2 pr-8 leading-tight bg-zinc-900 border border-zinc-400 rounded shadow appearance-none text-primary hover:border-zinc-500 focus:outline-none focus:ring"
-        >
+        <SelectTopic>
           <option value="">Buscar</option>
           {sortedTopics?.map((topic, index) => (
             <option
@@ -56,7 +54,8 @@ const MentorList: React.FC<MentorListProps> = ({ topics, children }) => {
               {topic.title}
             </option>
           ))}
-        </select>
+        </SelectTopic>
+
         <div className="absolute inset-y-0 right-0 flex items-center px-2 pointer-events-none text-primary">
           <svg
             className="w-4 h-4 fill-current"
@@ -70,7 +69,9 @@ const MentorList: React.FC<MentorListProps> = ({ topics, children }) => {
 
       <div className="flex flex-col min-h-screen align-center">
         <div className="grid grid-cols-1 gap-12 lg:grid-cols-2 auto-rows-min">
-          {children}
+          {filteredMentors.map((mentor, index) => (
+            <MentorCard key={index} mentor={mentor} topics={topics} />
+          ))}
         </div>
       </div>
     </div>
